Nvim: markdown inline comment highlighting workaround
Inline highlighting has for a while been broken with treesitter for me.
Disabling regex highlighting makes it significantly better, since it's
not doubling up on colours, but the root problem is that it marks inline
code as being comments. That doesn't seem right, but brighter comments
seem like a good idea anyway, so let's just make comments orange
(updated in my Gruvbox fork) and not italics. That way the code should
at least look better in Markdown until I sort out the root issue.

Added treesitter playground as it helps with debugging this stuff.